fix bugs and release code

+++ build.sh
@@ -1,6 +1,36 @@
 #!/bin/bash
+# Usage:
+# ./build.sh inputfile outputfile
+# Both args must be specified.
 
-[ -z "${CLONEZILLACD}" ] && CLONEZILLACD = clonezilla_with_virtualbox.iso
+INPUT_CLONEZILLA="$1"
+CLONEZILLACD="$2"
+if [ -z "${DOWNLOAD_CLONEZILLA_PATH}" ]
+then
+    DOWNLOAD_CLONEZILLA_PATH="https://sourceforge.net/projects/clonezilla/files/\
+clonezilla_live_alternative\
+/20200703-focal/clonezilla-live-20200703-focal-amd64.iso/download"
+fi
+
+[ "$(whoami)" != "root" ] && { echo "[ERR] must be root to proceed"; exit 1; }
+
+## @gn help_()
+## @brief Usage help
+## @note Markdown-ready.
+## @ingroup createInstaller
+
+help_() {
+
+echo "**Usage:**  "
+echo "./build.sh input.iso output.iso  "
+echo "  "
+echo "The input ISO file is a standard CloneZilla Debian distribution.  "
+echo "The output ISO file is the same distribution augmented with virtualbox guest additions.  "
+echo "**Both** arguments must be specified.  "
+echo "  "
+}
+
+
 [ -z "${VERBOSE}" ] && VERBOSE=false
 
 ## @fn add_guest_additions_to_clonezilla_iso()
@@ -28,7 +58,7 @@
 add_guest_additions_to_clonezilla_iso() {
 
     VMPATH=$PWD
-    
+
     bind_mount_clonezilla_iso
 
     cat > squashfs-root/update_clonezilla.sh << EOF
@@ -77,7 +107,12 @@ EOF
 
     # after exit now back under live/. Update linux kernel:
 
-    check_files squashfs-root/boot/vmlinuz squashfs-root/boot/initrd.img
+    if ! [ -f squashfs-root/boot/vmlinuz ] || ! [ -f  squashfs-root/boot/initrd.img ]
+    then
+        echo "[ERR] Could not find boot files"
+        exit 2
+    fi
+
     cp -vf --dereference squashfs-root/boot/vmlinuz vmlinuz
     cp -vf --dereference squashfs-root/boot/initrd.img  initrd.img
 
@@ -87,25 +122,43 @@ EOF
 
     # this first ISO image is a "save" one: from virtual disk to clonezilla
     # image
-        
+
     rm  -rf "mnt2/live/squashfs-root/"
 
     [ ! -f "mnt2/syslinux/isohdpfx.bin" ] \
         && cp -vf "clonezilla/syslinux/isohdpfx.bin" "mnt2/syslinux"
 
     xorriso -split_size 2047m -as mkisofs  \
-	    -isohybrid-mbr "$2/syslinux/isohdpfx.bin"  \
+	    -isohybrid-mbr "${VMPATH}/clonezilla/syslinux/isohdpfx.bin"  \
             -c syslinux/boot.cat   -b syslinux/isolinux.bin   -no-emul-boot \
             -boot-load-size 4   -boot-info-table   -eltorito-alt-boot  \
             -e boot/grub/efi.img \
             -no-emul-boot   -isohybrid-gpt-basdat   -o "${CLONEZILLACD}" mnt2
+
+    # cleaning up
+
+    if mountpoint mnt
+    then
+        if ! umount -l mnt
+        then
+            echo "[ERR] Could not unmount mnt."
+            echo "[ERR] Fatal. Exiting..."
+            exit 4
+        fi
+    fi
+    rm -rf mnt
+    umount -R -l mnt2
+    rm -rf mnt2
+    rm -rf ISOFILES
 }
 
 bind_mount_clonezilla_iso() {
 
-    fetch_process_clonezilla_iso
-
-    if_fails $? "[ERR] Could not fetch CloneZilla ISO file"
+    if ! process_clonezilla_iso
+    then
+        echo "[ERR] Could not fetch CloneZilla ISO file"
+        exit 1
+    fi
 
     local verb=""
     "${VERBOSE}" && verb="-v"
@@ -119,7 +172,7 @@ bind_mount_clonezilla_iso() {
         echo "[ERR] Could not create ISOFILES/home/partimag/"
         exit 1
     fi
-    
+
     if ! [ -d mnt2 ]
     then
         echo "[ERR] Could not create mnt2"
@@ -130,21 +183,21 @@ bind_mount_clonezilla_iso() {
         && echo "[INF] Now copying CloneZilla files to temporary \
 folder ISOFILES"
     rsync -a mnt2/ ISOFILES
-    
+
     if ! [ -d mnt2/syslinux ]
     then
         echo "[ERR] Could not create mnt2/syslinux"
         exit 1
     fi
-    
+
     if ! [ -f clonezilla/savedisk/isolinux.cfg ]
     then
         echo "[ERR] Could not find clonezilla/savedisk/isolinux.cfg"
         exit 1
     fi
-    
+
     cp ${verb} -f clonezilla/savedisk/isolinux.cfg mnt2/syslinux/
-        
+
     if ! [ -d mnt2/live ]
     then
         echo "[ERR] Could not create mnt2/live"
@@ -157,12 +210,12 @@ folder ISOFILES"
 
     for i in proc sys dev run; do mount -B /$i squashfs-root/$i; done
 
-    if [ $? != 0 ] 
-    then 
-        echo "[ERR] Could not bind-mount squashfs-root" 
+    if [ $? != 0 ]
+    then
+        echo "[ERR] Could not bind-mount squashfs-root"
         exit 2
     fi
-    
+
 }
 
 unmount_clonezilla_iso() {
@@ -170,12 +223,11 @@ unmount_clonezilla_iso() {
     # clean up and restore squashfs back
 
     rm -vf filesystem.squashfs
-    for i in proc sys dev run; do umount squashfs-root/$i; done
-    [ $? != 0] && echo "[ERR] Could not unmount squashfs-root"
-
+    for i in proc sys dev run; do umount -l squashfs-root/$i; done
+    [ $? != 0 ] && echo "[ERR] Could not unmount squashfs-root"
     mksquashfs squashfs-root filesystem.squashfs
     [ $? != 0 ] && echo "[ERR] Could not recreate squashfs filesystem"
-    
+
     cd "${VMPATH}"
 }
 
@@ -190,12 +242,37 @@ unmount_clonezilla_iso() {
 ## @retval 0 on success or exits -1 on failure.
 ## @ingroup fetchFunctions
 
-fetch_process_clonezilla_iso() {
+process_clonezilla_iso() {
 
     # std clonezilla iso is supposed to be in the root directory
     cd "${VMPATH}" || exit 2
     local verb=""
     "${VERBOSE}" && verb=-v
+
+    if [ -f "${INPUT_CLONEZILLA}" ]
+    then
+        echo "[MSG] Using ${INPUT_CLONEZILLA}..."
+    else
+      if ! curl -L "${DOWNLOAD_CLONEZILLA_PATH}" -o "${INPUT_CLONEZILLA}"
+      then
+        ${LOG[*]} "[ERR] Could not download CloneZilla CD from sourceforge."
+        exit 1
+      fi
+    fi
+
+    local md5
+    local md5_
+    local sha1
+    local sha1_
+    md5=$(md5sum "${INPUT_CLONEZILLA}"   | cut -d' ' -f 1)
+    md5_=$(cat SUMS.txt | grep MD5SUM    | cut -d' ' -f 2)
+    sha1=$(sha1sum "${INPUT_CLONEZILLA}" | cut -d' ' -f 1)
+    sha1_=$(cat SUMS.txt | grep SHA1SUM  | cut -d' ' -f 2)
+    if [ "${md5}" != "${md5_}" ] || [ "${sha1}" != "${sha1_}" ]
+    then
+        echo "[ERR] Checksums of ${INPUT_CLONEZILLA} and SUMS.txt do not match"
+        exit 4
+    fi
 
     # now cleanup, mount and copy CloneZilla live CD
 
@@ -217,15 +294,19 @@ fetch_process_clonezilla_iso() {
 
     [ ! -d mnt2 ] &&  mkdir mnt2  ||  { rm ${verb} -rf mnt2 && mkdir mnt2; }
 
-    "${VERBOSE}"  && echo "[INF] Mounting CloneZilla CD ${CLONEZILLACD}"
-    mount -oloop "${CLONEZILLACD}" ./mnt  \
-     	|| echo "[ERR] Could not mount ${CLONEZILLACD} to mnt"
+    "${VERBOSE}"  && echo "[INF] Mounting CloneZilla CD ${INPUT_CLONEZILLA}"
+
+    mount -oloop "${INPUT_CLONEZILLA}" ./mnt  \
+     	|| { echo "[ERR] Could not mount ${INPUT_CLONEZILLA} to mnt"
              exit 1; }
+
     "${VERBOSE}" \
         && echo "[INF] Now syncing CloneZilla CD to mnt2 in rw mode."
+
     rsync ${verb} -a ./mnt/ mnt2 \
-    	|| echo "[ERR] Could not copy clonezilla files to mnt2"
+    	|| { echo "[ERR] Could not copy clonezilla files to mnt2"
              exit 1; }
+
     cd mnt2/live || exit 2
     unsquashfs filesystem.squashfs \
       || { echo "[ERR] Failed to unsquash clonezilla's filesystem.squashfs"
@@ -234,3 +315,13 @@ fetch_process_clonezilla_iso() {
     cd "${VMPATH}"
     return 0
 }
+
+
+if [ $# -le 1 ]
+then
+    echo "[ERR] Both input and output arguments must be specified."
+    help_
+    exit 1
+else
+    add_guest_additions_to_clonezilla_iso
+fi
